# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for EchoHub Community Forum",
        "title": "EchoHub Community Forum API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/addNewPost": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Add a new post with title, content, and categories",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Posts"
                ],
                "summary": "Create post",
                "parameters": [
                    {
                        "description": "Post data",
                        "name": "newPost",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.PostData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post created successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/deletePost": {
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Remove an existing post by its ID",
                "tags": [
                    "Posts"
                ],
                "summary": "Delete post",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Post ID",
                        "name": "ID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Error"
                        }
                    }
                }
            }
        },
...
        "/newPost": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Render the page for creating a new forum post",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Posts"
                ],
                "summary": "Show new post page",
                "responses": {
                    "200": {
                        "description": "New post page HTML",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Error"
                        }
                    }
                }
            }
        },
...
        "/updatePost": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Render the page for editing an existing forum post",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Posts"
                ],
                "summary": "Show update post page",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Post ID",
                        "name": "ID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update post page HTML",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Error"
                        }
                    }
                }
            }
        },
...
        "/updatingPost": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Apply changes to title, content, or categories of a post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Posts"
                ],
                "summary": "Update post",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Post ID",
                        "name": "ID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Updated post data",
                        "name": "toUpdate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.PostData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "echohub-community_internal_models.Error": {
            "type": "object",
            "properties": {
                "Message": {
                    "type": "string"
                },
                "StatusCode": {
                    "type": "integer"
                },
                "SubMessage": {
                    "type": "string"
                },
                "Type": {
                    "type": "string"
                },
                "User": {
                    "$ref": "#/definitions/echohub-community_internal_models.User"
                }
            }
        },
...
        "echohub-community_internal_models.PostData": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "selectedCategories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "echohub-community_internal_models.Response": {
            "type": "object",
            "properties": {
                "extra": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "echohub-community_internal_models.User": {
            "type": "object",
            "properties": {
                "CreationDate": {
                    "type": "string"
                },
                "Email": {
                    "type": "string"
                },
                "Gender": {
                    "type": "string"
                },
                "ID": {
                    "type": "integer"
                },
                "ProfileImg": {
                    "type": "string"
                },
                "UserName": {
                    "type": "string"
                },
                "UserType": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "CookieAuth": {
            "description": "Session-based authentication using cookies",
            "type": "apiKey",
            "name": "userSession",
            "in": "cookie"
        }
    }
}